refs piwik/piwik#5248 initial work on the bandwidth plugin.
Turns out it will be most likely not doable this way by archiving
the bandwidth report and merging it with the getPageUrls eg because
of different subtable ids. Committing the current state of this
plugin tough in case we have to go this way. The next step would
be trying to make the actions plugin extensible (Archiving and API)